# docs/openapi/backend-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/calendar/events/{id}": {
      "patch": {
        "operationId": "CalendarController_updateEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      },
      "delete": {
        "operationId": "CalendarController_deleteEvent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
...
    "/calendar/integrations/{provider}/connect": {
      "post": {
        "operationId": "CalendarController_connectIntegration",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
    "/calendar/integrations/{provider}/oauth/start": {
      "post": {
        "operationId": "CalendarController_beginOauth",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
    "/calendar/integrations/{provider}/oauth/complete": {
      "post": {
        "operationId": "CalendarController_completeOauth",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
    "/calendar/integrations/{provider}/disconnect": {
      "post": {
        "operationId": "CalendarController_disconnectIntegration",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
    "/calendar/integrations/{provider}/sync": {
      "post": {
        "operationId": "CalendarController_syncIntegration",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Calendar"
        ]
      }
    },
...
    "/investments/posts/{id}/commitments": {
      "get": {
        "operationId": "InvestmentsController_getCommitments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Investments"
        ]
      },
      "post": {
        "operationId": "InvestmentsController_createCommitment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Investments"
        ]
      }
    },
...
    "/marketing/campaigns/{id}/status": {
      "post": {
        "operationId": "MarketingController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Marketing"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vendrome API",
    "description": "Vendrome backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}